Move theme builder composition root and remaining grid-specific components into ag-website-shared
Addresses feedback that the grid ThemeBuilder root (jotai store init, initial
preset, change-tracking wiring, Provider) was stuck in ag-grid-docs, forcing
AG Studio to duplicate this setup rather than reuse it.

- Extract Preset type + applyPreset() into ag-website-shared/theming/preset.ts
  (already fully generic, just stranded in grid's presets.ts)
- Extract the store/Provider bootstrap into a shared ThemeBuilderProvider
  component, parameterised by initialPreset + children
- Move ShadowDom.tsx (generic, no grid-specific logic) into shared
- Relocate the remaining grid-specific theme-builder tree into
  ag-website-shared/components/theme-builder-grid/, mirroring where a future
  Studio implementation would live
- Add a setThemeBuilderDocsUrl() configure-once hook so
  ThemeImportExportDialog no longer needs grid's urlWithBaseUrl directly
- Add a @generated-reference/* alias so registerParamDocs.ts's import of the
  generated theming-api.AUTO.json resolves regardless of the file's physical
  location
