+++ documentation/ag-grid-docs/src/pages/theme-builder.astro
@@ -1,9 +1,11 @@
 ---
-import { ThemeBuilder } from '@components/theme-builder/ThemeBuilder';
 import { getCollection } from 'astro:content';
+
+import { EmptyState } from '@ag-website-shared/components/theme-builder-grid/EmptyState';
+import { ThemeBuilder } from '@ag-website-shared/components/theme-builder-grid/ThemeBuilder';
+
 import { getDocsPages } from '../components/docs/utils/pageData';
 import Layout from '../layouts/Layout.astro';
-import { EmptyState } from '@components/theme-builder/EmptyState';
 
 export async function getStaticPaths() {
     const pages = await getCollection('docs');
