+++ external/ag-website-shared/src/components/ShadowDom.tsx
@@ -7,7 +7,6 @@ export type ShadowDomProps = {
 
 export const ShadowDom = ({ children }: ShadowDomProps) => {
     const [isHydrated, setIsHydrated] = useState(false);
-    // eslint-disable-next-line react-hooks/set-state-in-effect -- intentional hydration pattern
     useEffect(() => setIsHydrated(true), []);
 
     return <>{isHydrated && <root.div style={{ height: '100%' }}>{children}</root.div>}</>;
