+++ external/ag-website-shared/src/components/theme-builder-grid/components/grid-config/grid-config-atom.ts
@@ -18,7 +18,6 @@ export const useGridOptions = () => {
     const gridOptions = useMemo(() => {
         return buildGridOptions(config);
     }, [config]);
-    /* eslint-disable react-hooks/refs -- intentionally reading/writing ref during render to track config changes */
     const state = useRef({ updateCount: 1, prevConfig: config });
     if (config !== state.current.prevConfig) {
         state.current.updateCount += 1;
@@ -29,5 +28,4 @@ export const useGridOptions = () => {
         config,
         updateCount: state.current.updateCount,
     };
-    /* eslint-enable react-hooks/refs */
 };
