+++ external/ag-website-shared/src/components/theme-builder-grid/model/registerParamDocs.ts
@@ -1,5 +1,4 @@
 import { setParamDocsProvider } from '@ag-website-shared/theming/ParamModel';
-
-import themingApi from '../../../../../../dist/documentation/reference/theming-api.AUTO.json';
+import themingApi from '@generated-reference/theming-api.AUTO.json';
 
 setParamDocsProvider((param) => (themingApi as Record<string, { meta?: { comment?: string } }>)[param]?.meta?.comment);
