+++ external/ag-website-shared/src/components/theme-builder-grid/registerEditorConfig.ts
@@ -1,5 +1,7 @@
+import { setThemeBuilderDocsUrl } from '@ag-website-shared/components/theme-builder-grid/components/general/ThemeImportExportDialog';
 import { setFontFamilyOptions } from '@ag-website-shared/components/theme-builder/FontFamilyValueEditor';
 import { setImageValuesDocsUrl } from '@ag-website-shared/components/theme-builder/ImageValueEditor';
+import { urlWithBaseUrl } from '@utils/urlWithBaseUrl';
 
 setFontFamilyOptions([
     {
@@ -67,3 +69,5 @@ setFontFamilyOptions([
 ]);
 
 setImageValuesDocsUrl('/react-data-grid/theming-parameters/#image-values');
+
+setThemeBuilderDocsUrl(urlWithBaseUrl('/data-grid/theming-theme-builder/'));
